@@ -13,10 +13,10 @@
     "formik": "2.2.9",
     "gh-pages": "3.2.3",
     "jest": "27.5.1",
-    "react": "17.0.2",
+    "react": "^18.2.0",
     "react-dom": "17.0.2",
     "react-redux": "7.2.6",
-    "react-router-dom": "5.3.0",
+    "react-router-dom": "^6.5.0",
     "react-scripts": "5.0.0",
     "react-test-renderer": "17.0.2",
     "redux": "4.1.2",
@@ -58,7 +58,7 @@
     "@types/react": "17.0.20",
     "@types/react-dom": "17.0.9",
     "@types/react-redux": "7.1.22",
-    "@types/react-router-dom": "5.3.3",
+    "@types/react-router-dom": "^5.3.3",
     "@types/react-test-renderer": "17.0.1",
     "@types/redux-form": "8.3.3",
     "@types/uuid": "8.3.4"
