@@ -6,24 +6,10 @@ import {Message} from "./Message/MessageItem";
 
 export const Dialogs = (props) => {
 
-    let dialogs = [
-        {id: 1, name: 'Dimych'},
-        {id: 2, name: 'Andrey'},
-        {id: 3, name: 'Sveta'},
-        {id: 4, name: 'Sasha'},
-        {id: 5, name: 'Viktor'},
-        {id: 6, name: 'Valera'},
-    ]
 
-    let messages = [
-        {id: 1, message: 'Hi'},
-        {id: 2, message: 'How is your it-kamasutra'},
-        {id: 3, message: 'Yo'},
-    ]
+    let dialogElements = props.state.dialogs.map(d => <DialogItem name={d.name} id={d.id}/>)
 
-    let dialogElements = dialogs.map(d => <DialogItem name={d.name} id={d.id}/>)
-
-    let messageElement = messages.map(m => <Message message={m.message} id={m.id}/>)
+    let messageElement = props.state.messages.map(m => <Message message={m.message} id={m.id}/>)
 
 
     return (
