@@ -2,8 +2,11 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import {App} from './App';
+import {state} from "./redux/State";
+
 
 ReactDOM.render(
-    <App />,
-  document.getElementById('root')
+    <App state={state}
+    />,
+    document.getElementById('root')
 );
